#5 [feat/#5-middleware] - guard middleware로 인증된 사용자인지 체크
인증이 필요한 부분 검열

controller 로직 부분에 req.user로 넘겨온게 본인이랑 일치하는지 확인

+++ server1/controllers/question.controller.js
@@ -6,13 +6,12 @@ class QuestionController {
     try {
       const { title, content } = req.body;
       /**TODO: uid --change--> uid in parsing guard middleware */
-      const { uid } = req.query;
+      const { uid } = req.user;
       const newQuestion = new Question(uid, title, content);
       questionService.create(newQuestion);
       res.status(201).json({ message: "Question created successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
   async read(req, res, next) {
@@ -20,31 +19,31 @@ class QuestionController {
       const questions = await questionService.read();
       res.json(questions);
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
   async update(req, res, next) {
     try {
-      const { uid } = req.query;
-      const { qid, title, content } = req.body;
+      const { uid, qid } = req.query;
+      const { title, content } = req.body;
       const updatedQuestion = new Question(uid, title, content);
+      if (req.user.uid !== uid) throw new Error("not Matched User");
+
       questionService.update(qid, updatedQuestion);
       res.status(201).json({ message: "Question updated successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 
   async delete(req, res, next) {
     try {
-      const { qid } = req.query;
+      const { qid, uid } = req.query;
+      if (req.user.uid !== uid) throw new Error("not Matched User");
       questionService.delete(qid);
       res.status(201).json({ message: "Question deleted successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 }
