+++ server1/controllers/user.controller.js
@@ -9,8 +9,7 @@ class UserController {
       userService.create(newUser);
       res.status(201).json({ message: "User created successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 
@@ -19,8 +18,7 @@ class UserController {
       const users = await userService.read();
       res.json(users);
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 
@@ -29,22 +27,24 @@ class UserController {
       const { uid } = req.query;
       const { username, email, password, region, tel } = req.body;
       const newUser = new User(username, email, password, region, tel);
+
+      if (req.user.uid !== uid) throw new Error("not Matched User");
       userService.update({ uid: uid, ...newUser });
       res.status(201).json({ message: "User updated successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 
   async delete(req, res, next) {
     try {
       const { uid } = req.query;
+      if (req.user.uid !== uid) throw new Error("not Matched User");
+
       userService.delete(uid);
       res.status(201).json({ message: "User deleted successfully." });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Internal server error." });
+      next(error);
     }
   }
 }
